+++ src/main/java/com/discordbolt/boltbot/discord/system/sync/GuildJoinListener.java
@@ -1,8 +1,8 @@
 package com.discordbolt.boltbot.discord.system.sync;
 
-import com.discordbolt.boltbot.data.objects.GuildData;
-import com.discordbolt.boltbot.data.repositories.GuildRepository;
 import com.discordbolt.boltbot.discord.EventListener;
+import com.discordbolt.boltbot.repository.GuildRepository;
+import com.discordbolt.boltbot.repository.entity.GuildData;
 import discord4j.core.event.domain.guild.GuildCreateEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
