+++ src/main/java/com/discordbolt/boltbot/discord/system/sync/UserJoinListener.java
@@ -1,8 +1,8 @@
 package com.discordbolt.boltbot.discord.system.sync;
 
-import com.discordbolt.boltbot.data.objects.UserData;
-import com.discordbolt.boltbot.data.repositories.UserRepository;
 import com.discordbolt.boltbot.discord.EventListener;
+import com.discordbolt.boltbot.repository.UserRepository;
+import com.discordbolt.boltbot.repository.entity.UserData;
 import discord4j.core.event.domain.guild.MemberJoinEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
