+++ src/main/java/com/discordbolt/boltbot/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package com.discordbolt.boltbot.data.repositories;
+package com.discordbolt.boltbot.repository;
 
-import com.discordbolt.boltbot.data.objects.GuildData;
-import com.discordbolt.boltbot.data.objects.UserData;
+import com.discordbolt.boltbot.repository.entity.GuildData;
+import com.discordbolt.boltbot.repository.entity.UserData;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
