+++ src/main/java/com/discordbolt/boltbot/web/controllers/IndexController.java
@@ -1,12 +1,14 @@
-package com.discordbolt.boltbot.web;
+package com.discordbolt.boltbot.web.controllers;
 
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping(path = "/")
 public class IndexController {
 
-    @RequestMapping("/")
+    @GetMapping(path = "")
     public String index() {
         return "Greetings from Spring Boot!";
     }
