+++ src/main/java/com/discordbolt/boltbot/web/exceptions/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package com.discordbolt.boltbot.web.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class EntityNotFoundException extends RuntimeException {
+}
